requirements.txt
streamlit
groq
python-dotenv
chromadb
sentence-transformers
PyPDF2
easyocr
